+++ src/main/java/com/supermartijn642/movingelevators/elevator/ElevatorGroupCapability.java
@@ -2,6 +2,7 @@ package com.supermartijn642.movingelevators.elevator;
 
 import com.supermartijn642.movingelevators.MovingElevators;
 import com.supermartijn642.movingelevators.blocks.ControllerBlockEntity;
+import com.supermartijn642.movingelevators.extensions.MovingElevatorsLevel;
 import com.supermartijn642.movingelevators.packets.PacketAddElevatorGroup;
 import com.supermartijn642.movingelevators.packets.PacketRemoveElevatorGroup;
 import com.supermartijn642.movingelevators.packets.PacketUpdateElevatorGroups;
@@ -9,93 +10,45 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.Tag;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
-import net.minecraftforge.common.capabilities.*;
-import net.minecraftforge.common.util.LazyOptional;
-import net.minecraftforge.event.AttachCapabilitiesEvent;
-import net.minecraftforge.event.TickEvent;
-import net.minecraftforge.event.entity.player.PlayerEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
+import net.neoforged.neoforge.common.NeoForge;
+import net.neoforged.neoforge.event.TickEvent;
+import net.neoforged.neoforge.event.entity.player.PlayerEvent;
 
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.function.Consumer;
 
 /**
  * Created 11/7/2020 by SuperMartijn642
  */
-@Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class ElevatorGroupCapability {
 
-    private static Capability<ElevatorGroupCapability> CAPABILITY = CapabilityManager.get(new CapabilityToken<ElevatorGroupCapability>() {
-    });
-
     public static ElevatorGroupCapability get(Level level){
-        return level.getCapability(CAPABILITY).orElse(null);
+        return ((MovingElevatorsLevel)level).getElevatorGroupCapability();
     }
 
-    @SubscribeEvent
-    public static void register(RegisterCapabilitiesEvent e){
-        e.register(ElevatorGroupCapability.class);
-    }
-
-    @SubscribeEvent
-    public static void attachCapabilities(AttachCapabilitiesEvent<Level> e){
-        Level level = e.getObject();
-
-        LazyOptional<ElevatorGroupCapability> capability = LazyOptional.of(() -> new ElevatorGroupCapability(level));
-        e.addCapability(new ResourceLocation("movingelevators", "elevator_groups"), new ICapabilitySerializable<Tag>() {
-            @Nonnull
-            @Override
-            public <T> LazyOptional<T> getCapability(@Nonnull Capability<T> cap, @Nullable Direction side){
-                return cap == CAPABILITY ? capability.cast() : LazyOptional.empty();
-            }
-
-            @Override
-            public Tag serializeNBT(){
-                return capability.map(ElevatorGroupCapability::write).orElse(null);
-            }
-
-            @Override
-            public void deserializeNBT(Tag nbt){
-                capability.ifPresent(capability -> capability.read(nbt));
-            }
+    public static void registerEventListeners(){
+        NeoForge.EVENT_BUS.addListener((Consumer<TickEvent.LevelTickEvent>)event -> {
+            if(event.side.isServer() && event.phase == TickEvent.Phase.END)
+                tickWorldCapability(event.level);
         });
-        e.addListener(capability::invalidate);
-    }
-
-
-    @SubscribeEvent
-    public static void onTick(TickEvent.LevelTickEvent e){
-        if(e.level.isClientSide || e.phase != TickEvent.Phase.END)
-            return;
-
-        tickWorldCapability(e.level);
+        NeoForge.EVENT_BUS.addListener((Consumer<PlayerEvent.PlayerChangedDimensionEvent>)event -> onJoinWorld(event.getEntity(), event.getEntity().level()));
+        NeoForge.EVENT_BUS.addListener((Consumer<PlayerEvent.PlayerLoggedInEvent>)event -> onJoin(event.getEntity()));
     }
 
     public static void tickWorldCapability(Level level){
-        level.getCapability(CAPABILITY).ifPresent(ElevatorGroupCapability::tick);
+        get(level).tick();
     }
 
-    @SubscribeEvent
-    public static void onJoinWorld(PlayerEvent.PlayerChangedDimensionEvent e){
-        ServerPlayer player = (ServerPlayer)e.getEntity();
-        player.level().getCapability(CAPABILITY).ifPresent(groups ->
-            MovingElevators.CHANNEL.sendToPlayer(player, new PacketUpdateElevatorGroups(groups.write()))
-        );
+    public static void onJoinWorld(Player player, Level level){
+        MovingElevators.CHANNEL.sendToPlayer(player, new PacketUpdateElevatorGroups(get(level).write()));
     }
 
-    @SubscribeEvent
-    public static void onJoin(PlayerEvent.PlayerLoggedInEvent e){
-        ServerPlayer player = (ServerPlayer)e.getEntity();
-        player.level().getCapability(CAPABILITY).ifPresent(groups ->
-            MovingElevators.CHANNEL.sendToPlayer(player, new PacketUpdateElevatorGroups(groups.write()))
-        );
+    public static void onJoin(Player player){
+        MovingElevators.CHANNEL.sendToPlayer(player, new PacketUpdateElevatorGroups(get(player.level()).write()));
     }
 
     private final Level level;