+++ src/main/java/com/supermartijn642/movingelevators/extensions/MovingElevatorsLevel.java
@@ -1,0 +1,11 @@
+package com.supermartijn642.movingelevators.extensions;
+
+import com.supermartijn642.movingelevators.elevator.ElevatorGroupCapability;
+
+/**
+ * Created 08/03/2023 by SuperMartijn642
+ */
+public interface MovingElevatorsLevel {
+
+    ElevatorGroupCapability getElevatorGroupCapability();
+}