+++ src/main/java/com/supermartijn642/movingelevators/mixin/ServerLevelMixin.java
@@ -1,0 +1,38 @@
+package com.supermartijn642.movingelevators.mixin;
+
+import com.supermartijn642.movingelevators.elevator.ElevatorGroupCapability;
+import com.supermartijn642.movingelevators.elevator.ElevatorGroupCapabilitySaveData;
+import net.minecraft.resources.ResourceKey;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.progress.ChunkProgressListener;
+import net.minecraft.world.RandomSequences;
+import net.minecraft.world.level.CustomSpawner;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.dimension.LevelStem;
+import net.minecraft.world.level.storage.LevelStorageSource;
+import net.minecraft.world.level.storage.ServerLevelData;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import java.util.List;
+import java.util.concurrent.Executor;
+
+/**
+ * Created 21/02/2023 by SuperMartijn642
+ */
+@Mixin(ServerLevel.class)
+public class ServerLevelMixin {
+
+    @Inject(
+        method = "<init>",
+        at = @At("TAIL")
+    )
+    private void constructor(MinecraftServer minecraftServer, Executor executor, LevelStorageSource.LevelStorageAccess levelStorageAccess, ServerLevelData serverLevelData, ResourceKey<Level> resourceKey, LevelStem levelStem, ChunkProgressListener chunkProgressListener, boolean bl, long l, List<CustomSpawner> list, boolean bl2, RandomSequences randomSequences, CallbackInfo ci){
+        //noinspection DataFlowIssue
+        ServerLevel level = (ServerLevel)(Object)this;
+        ElevatorGroupCapabilitySaveData.init(level, ElevatorGroupCapability.get(level));
+    }
+}