+++ changelog.md
@@ -1,2 +1,2 @@
 ### Moving Elevators 1.4.5
-- Initial release of Moving Elevators for Minecraft 1.20.3 & 1.20.4
+- Initial release of Moving Elevators for NeoForge